Update names and description with new field
