+++ .github/workflows/ci.yml
@@ -51,7 +51,7 @@ jobs:
         with:
           command: fetch
       - name: Build for target
-        run: cargo build --verbose --no-default-features --features f63 --target ${{ matrix.target }}
+        run: cargo build --verbose --no-default-features --features f64 --target ${{ matrix.target }}
 
   bitrot:
     name: Bitrot check
