+++ src/rescue_64_14_7/digest.rs
