+++ src/rescue_64_8_4/hasher.rs
@@ -84,20 +84,20 @@ impl RescueHash {
     }
 
     /// Serializes the current state to an array of bytes
-    pub fn to_bytes(&self) -> [u8; 120] {
-        let mut res = [0u8; 120];
+    pub fn to_bytes(&self) -> [u8; 72] {
+        let mut res = [0u8; 72];
         assert_eq!(res.len(), STATE_WIDTH * 8 + 8);
 
         for (index, elem) in self.state.iter().enumerate() {
             res[index * 8..index * 8 + 8].copy_from_slice(&elem.to_bytes());
         }
-        res[112..120].copy_from_slice(&(self.idx as u64).to_le_bytes());
+        res[64..72].copy_from_slice(&(self.idx as u64).to_le_bytes());
 
         res
     }
 
     /// Returns a RescueHash from an array of bytes
-    pub fn from_bytes(bytes: &[u8; 120]) -> Result<Self, SerializationError> {
+    pub fn from_bytes(bytes: &[u8; 72]) -> Result<Self, SerializationError> {
         let mut state = [Fp::zero(); STATE_WIDTH];
         let mut array = [0u8; 8];
         for index in 0..STATE_WIDTH {
@@ -109,7 +109,7 @@ impl RescueHash {
             };
         }
 
-        array.copy_from_slice(&bytes[112..120]);
+        array.copy_from_slice(&bytes[64..72]);
         let idx = u64::from_le_bytes(array) as usize;
 
         Ok(Self { state, idx })
@@ -207,7 +207,6 @@ impl Hasher for RescueHash {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use cheetah::group::ff::Field;
     use rand::thread_rng;
 
     #[test]
@@ -216,209 +215,143 @@ mod tests {
         // Sagemath code at https://github.com/KULeuven-COSIC/Marvellous
 
         let input_data = [
-            [Fp::zero(); 7],
-            [Fp::one(); 7],
+            [Fp::zero(); 4],
+            [Fp::one(); 4],
             [
-                Fp::new(3950023656837154780),
-                Fp::new(2881042141171700070),
-                Fp::new(10779783730324926223),
-                Fp::new(2193245569390396079),
-                Fp::new(16253948482548340315),
-                Fp::new(1218876753142647280),
-                Fp::new(4665677124148004089),
+                Fp::new(12696789723516079038),
+                Fp::new(9199133001420111383),
+                Fp::new(4917625779728435204),
+                Fp::new(1296807772188121589),
             ],
             [
-                Fp::new(15510680881292026648),
-                Fp::new(1885599458732076372),
-                Fp::new(3009832769468343331),
-                Fp::new(4278715578773017787),
-                Fp::new(1527318565755464029),
-                Fp::new(15305488897400747380),
-                Fp::new(16465990826098742194),
+                Fp::new(6576572786004571242),
+                Fp::new(13520483611514881578),
+                Fp::new(7396734565988624997),
+                Fp::new(3797367628558919976),
             ],
             [
-                Fp::new(140365944095159147),
-                Fp::new(15974833700665793655),
-                Fp::new(4537294998186101657),
-                Fp::new(5109852931603194239),
-                Fp::new(17928282694277838009),
-                Fp::new(13253300241959402999),
-                Fp::new(3823973534601399165),
+                Fp::new(12711665381750550530),
+                Fp::new(3085138097114110958),
+                Fp::new(13276586898730573338),
+                Fp::new(2170068693998401624),
             ],
             [
-                Fp::new(17954336847934734809),
-                Fp::new(12987799873114840312),
-                Fp::new(12918391975014059030),
-                Fp::new(13534406757215195217),
-                Fp::new(7015568714821117231),
-                Fp::new(16406332556186741701),
-                Fp::new(13841094173526572577),
+                Fp::new(1083726018176650923),
+                Fp::new(5602938554164977237),
+                Fp::new(9503302027124828018),
+                Fp::new(17851032627776582496),
             ],
             [
-                Fp::new(5639224153506147889),
-                Fp::new(11045864791932293505),
-                Fp::new(16971439160821907715),
-                Fp::new(11823824893550570395),
-                Fp::new(6530760332296797048),
-                Fp::new(16658398244005217282),
-                Fp::new(17596930911850813020),
+                Fp::new(3460901845865360280),
+                Fp::new(8874900455910869977),
+                Fp::new(12961454695644130877),
+                Fp::new(15224475753097411894),
             ],
             [
-                Fp::new(14177301956587338153),
-                Fp::new(11152794082924817567),
-                Fp::new(6237926186494600489),
-                Fp::new(2860546146811461105),
-                Fp::new(1945982128333621329),
-                Fp::new(16175721111896472344),
-                Fp::new(12470940840517216660),
+                Fp::new(7751822847244044845),
+                Fp::new(14439542176698867963),
+                Fp::new(2452137216376559421),
+                Fp::new(16783572658585168490),
             ],
             [
-                Fp::new(4108451947597787380),
-                Fp::new(17716249113802510600),
-                Fp::new(2842886041996649217),
-                Fp::new(3125385885849793569),
-                Fp::new(14222444035786103231),
-                Fp::new(14291592125380804622),
-                Fp::new(15788476122562488600),
+                Fp::new(788300095521646318),
+                Fp::new(13480310207084563773),
+                Fp::new(5432697520099597064),
+                Fp::new(5640374229052330762),
             ],
             [
-                Fp::new(6324039691650518534),
-                Fp::new(4854235111930838947),
-                Fp::new(9528543306599103227),
-                Fp::new(1908154389856780779),
-                Fp::new(725275410145065933),
-                Fp::new(2994053441956803602),
-                Fp::new(14870768704779351904),
+                Fp::new(9585869648207771849),
+                Fp::new(2964532145699725522),
+                Fp::new(5111097433776572204),
+                Fp::new(3948243718771775964),
             ],
             [
-                Fp::new(16708718752419310323),
-                Fp::new(9043524975036485217),
-                Fp::new(10559977491562365190),
-                Fp::new(10805193695687328688),
-                Fp::new(3342433955286461953),
-                Fp::new(6763271737144514490),
-                Fp::new(8128474802178971483),
+                Fp::new(7558703435766799826),
+                Fp::new(6807316245062936198),
+                Fp::new(1670996120415067330),
+                Fp::new(2653701437900945479),
             ],
             [
-                Fp::new(14371251217142520515),
-                Fp::new(4338524858237943279),
-                Fp::new(17283938505215119784),
-                Fp::new(18443919525318776756),
-                Fp::new(11554353940375106506),
-                Fp::new(9935875077770851674),
-                Fp::new(16651331719620526423),
+                Fp::new(1869140707987786340),
+                Fp::new(7695687736955165162),
+                Fp::new(4384620636766221999),
+                Fp::new(11237537366181283950),
             ],
         ];
 
         // Generated from https://github.com/KULeuven-COSIC/Marvellous
         let output_data = [
             [
-                Fp::new(1462852121402184757),
-                Fp::new(3851389896230401122),
-                Fp::new(5882996393075625557),
-                Fp::new(10521291612941708615),
-                Fp::new(17822059276976522025),
-                Fp::new(6442736186431050368),
-                Fp::new(13017287424191436946),
+                Fp::new(4910989335886459515),
+                Fp::new(2796690540326658613),
+                Fp::new(13392979779619727901),
+                Fp::new(12220694230377879406),
             ],
             [
-                Fp::new(2775776797579705596),
-                Fp::new(2403614533261838082),
-                Fp::new(7693169250485223950),
-                Fp::new(11055303056864038887),
-                Fp::new(7628852160479342589),
-                Fp::new(12233463306084900855),
-                Fp::new(11936639582709302076),
+                Fp::new(14030948176935373137),
+                Fp::new(6889229919436197380),
+                Fp::new(15585020268064615960),
+                Fp::new(15230679386058804747),
             ],
             [
-                Fp::new(11505824700477102572),
-                Fp::new(16374638205843913216),
-                Fp::new(4248359725379407058),
-                Fp::new(6377440333976445055),
-                Fp::new(12153834900075533870),
-                Fp::new(16552688606391487587),
-                Fp::new(13487106855369955288),
+                Fp::new(706470175004787691),
+                Fp::new(9045846419159183300),
+                Fp::new(2676418206168607609),
+                Fp::new(4286167598147620515),
             ],
             [
-                Fp::new(16831576602944861321),
-                Fp::new(10554228133044132029),
-                Fp::new(7341467356168262539),
-                Fp::new(18048235872609925908),
-                Fp::new(15349705420822053264),
-                Fp::new(5099126063801993907),
-                Fp::new(15758960528207547734),
+                Fp::new(18380410717835921573),
+                Fp::new(12659766701504815380),
+                Fp::new(13863416848783506403),
+                Fp::new(4517319948670814325),
             ],
             [
-                Fp::new(6140546130285181457),
-                Fp::new(9717430031239952053),
-                Fp::new(426384473071363511),
-                Fp::new(8738424094168606662),
-                Fp::new(13895934231411749553),
-                Fp::new(9745694210096996441),
-                Fp::new(238205385641180480),
+                Fp::new(3457611930991995171),
+                Fp::new(15939660134846686812),
+                Fp::new(3930594105400200418),
+                Fp::new(8901001841441354026),
             ],
             [
-                Fp::new(12072258341513292338),
-                Fp::new(1653787536644408066),
-                Fp::new(2553325559739182980),
-                Fp::new(14494640366770314229),
-                Fp::new(6632328271058782797),
-                Fp::new(17136655182521628181),
-                Fp::new(9648599993263888708),
+                Fp::new(2174974189260558415),
+                Fp::new(674423569210728397),
+                Fp::new(14243685749685816500),
+                Fp::new(16221541406966092001),
             ],
             [
-                Fp::new(16172310988415901282),
-                Fp::new(3867868964145925742),
-                Fp::new(16640854393851664160),
-                Fp::new(13380412620490202361),
-                Fp::new(9755980514321283475),
-                Fp::new(17079300832339333147),
-                Fp::new(3148591176626069201),
+                Fp::new(3979697094498344164),
+                Fp::new(3871436604718095236),
+                Fp::new(17564369428971590635),
+                Fp::new(16368117003750996090),
             ],
             [
-                Fp::new(5495083559674337780),
-                Fp::new(12519343593850221554),
-                Fp::new(14454522298770281509),
-                Fp::new(7282539795336546358),
-                Fp::new(12637748394059553436),
-                Fp::new(3660318624754655797),
-                Fp::new(16464209702200801348),
+                Fp::new(14127398247547518574),
+                Fp::new(14195414343668611852),
+                Fp::new(2745986940484050712),
+                Fp::new(8510168344648313936),
             ],
             [
-                Fp::new(561228919855163821),
-                Fp::new(5745477782921558501),
-                Fp::new(14345931443850607969),
-                Fp::new(10911934545965820926),
-                Fp::new(11564909789117272275),
-                Fp::new(6575963491016418381),
-                Fp::new(6688464136460516678),
+                Fp::new(12357018050803044904),
+                Fp::new(12200585702753246383),
+                Fp::new(11782727060576301320),
+                Fp::new(7544131744653844429),
             ],
             [
-                Fp::new(5430170233120162060),
-                Fp::new(14073448356996281917),
-                Fp::new(9529268804005242596),
-                Fp::new(15835814726539502651),
-                Fp::new(18114992307639882028),
-                Fp::new(13124714119438319363),
-                Fp::new(17208402121800201187),
+                Fp::new(17366033595832430932),
+                Fp::new(17633758431123235206),
+                Fp::new(10255355719890378928),
+                Fp::new(1096989440083480883),
             ],
             [
-                Fp::new(502659996860556413),
-                Fp::new(7249969696684244242),
-                Fp::new(3655199103889939873),
-                Fp::new(6630626149995182382),
-                Fp::new(7831937114322467287),
-                Fp::new(4499093287722173119),
-                Fp::new(360388548637633151),
+                Fp::new(13019757464280815551),
+                Fp::new(1980740933937182560),
+                Fp::new(9449350348385637444),
+                Fp::new(12140448948176230382),
             ],
             [
-                Fp::new(9679775309692013648),
-                Fp::new(14664355940396898404),
-                Fp::new(10840083312362647488),
-                Fp::new(10643787539329357191),
-                Fp::new(4579477574868391271),
-                Fp::new(7919466403874804166),
-                Fp::new(549103385692248750),
+                Fp::new(18292577592561729440),
+                Fp::new(7683128746905765130),
+                Fp::new(2870487906647554545),
+                Fp::new(7565413129633922318),
             ],
         ];
 
@@ -458,7 +391,7 @@ mod tests {
         let mut hasher = RescueHash::new();
         hasher.update(&data);
 
-        let bytes = [255u8; 120];
+        let bytes = [255u8; 72];
 
         assert!(RescueHash::from_bytes(&bytes).is_err());
     }
