+++ src/rescue_64_8_4/sbox.rs
